+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/detail/DetailPlanScreen.kt
@@ -1,6 +1,5 @@
-package com.yapp.growth.presentation.ui.main.detailPlan
+package com.yapp.growth.presentation.ui.main.detail
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
@@ -14,17 +13,14 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.yapp.growth.presentation.R
 import com.yapp.growth.presentation.component.PlanzExitAppBar
 import com.yapp.growth.presentation.theme.*
-import com.yapp.growth.presentation.ui.main.detailPlan.DetailPlanContract.*
+import com.yapp.growth.presentation.ui.main.detail.DetailPlanContract.*
 import kotlinx.coroutines.flow.collect
 
 @Composable
